Extract SslCtx methods for better sharing
